refactor: move command playback scenarios
