+++ src/loushang/coding/ui/playback_scenarios/__init__.py
@@ -1,0 +1,2 @@
+"""Grouped native TUI playback scenario definitions."""
+
